commit after adding resent otp feature

+++ src/app/auth/components/user-signup/user-signup.component.ts
@@ -17,13 +17,18 @@ export class UserSignupComponent {
     
     this.userService.userSignup(datas).subscribe(
       (res)=>{
-        localStorage.setItem("signupData", JSON.stringify(datas));
+        if(res=='email already used'){
+          alert('this email is already used')
+        }else if(res=='status pending otp sent'){
+          localStorage.setItem("signupData", JSON.stringify(datas));
+          this.router.navigate(['/auth/SignupOtp'])
+        }else if(res=='incorrect email or password'){
+              alert('incorrect email or password')
+        }else{
 
-        this.router.navigate(['/auth/SignupOtp'], {
-          queryParams: {
-            option: 'signup'
-          }
-        })
+          localStorage.setItem("signupData", JSON.stringify(datas));
+          this.router.navigate(['/auth/SignupOtp'])
+        }
         
       },
       (err:any)=>{
